+++ packages/client/test/unit/Decrypt.test.ts
@@ -5,7 +5,7 @@ import { DestroySignal } from '../../src/DestroySignal'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { StreamrClientEventEmitter } from '../../src/events'
 import { Decrypt } from '../../src/subscribe/Decrypt'
-import { createMockMessage, mockContext } from '../test-utils/utils'
+import { createMockMessage, mockLoggerFactory } from '../test-utils/utils'
 
 describe('Decrypt', () => {
 
@@ -16,15 +16,14 @@ describe('Decrypt', () => {
         const keyExchange = {
             requestGroupKey: async () => {}
         }
-        const context = mockContext()
         const decrypt = new Decrypt(
-            context,
             groupKeyStore as any,
             keyExchange as any,
             {
                 clearStream: jest.fn()
             } as any,
-            new DestroySignal(context),
+            new DestroySignal(),
+            mockLoggerFactory(),
             new StreamrClientEventEmitter(),
             {
                 keyRequestTimeout: 50