+++ packages/client/test/unit/Validator.test.ts
@@ -5,7 +5,7 @@ import { Stream } from '../../src/Stream'
 import { StreamRegistry } from '../../src/registry/StreamRegistry'
 import { StreamRegistryCached } from '../../src/registry/StreamRegistryCached'
 import { Validator } from '../../src/Validator'
-import { createMockMessage, mockContext } from '../test-utils/utils'
+import { createMockMessage, mockLoggerFactory } from '../test-utils/utils'
 import { fastWallet } from 'streamr-test-utils'
 import { EthereumAddress } from '@streamr/utils'
 
@@ -23,10 +23,8 @@ const createMockValidator = () => {
             return userAddress.toLowerCase() === publisherWallet.address.toLowerCase()
         }
     }
-    const context = mockContext()
     return new Validator(
-        context,
-        new StreamRegistryCached(context, streamRegistry as any, {} as any) as any,
+        new StreamRegistryCached(mockLoggerFactory(), streamRegistry as any, {} as any) as any,
         {} as any
     )
 }
@@ -81,3 +79,4 @@ describe('Validator', () => {
         })
     })
 })
+